refactor: 型定義を mock-data から分離し、テスト追加・UI 改善
- Badge/MockGroup の型を types.ts に分離し、モック削除時の影響を局所化
- formatFanDuration のユニットテストを追加（7ケース）
- ProfileProgress の percentageIncrease が負の場合に text-error で表示
- ProfileGroups の画像に loading="lazy" を追加
